+++ pkg/local_object_storage/engine/shards.go
@@ -49,6 +49,14 @@ func (m *metricsWithID) SetReadonly(readonly bool) {
 	m.mw.SetReadonly(m.id, readonly)
 }
 
+func (m *metricsWithID) AddToContainerSize(cnr string, size int64) {
+	m.mw.AddToContainerSize(cnr, size)
+}
+
+func (m *metricsWithID) AddToPayloadSize(size int64) {
+	m.mw.AddToPayloadCounter(m.id, size)
+}
+
 // AddShard adds a new shard to the storage engine.
 //
 // Returns any error encountered that did not allow adding a shard.