+++ pkg/local_object_storage/shard/control.go
@@ -132,7 +132,7 @@ func (s *Shard) Init() error {
 		}
 	}
 
-	s.updateObjectCounter()
+	s.initMetrics()
 
 	s.gc = &gc{
 		gcCfg:       &s.gcCfg,